requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
jinja2==3.1.2
python-dotenv==1.0.0
gspread==5.12.0
oauth2client==4.1.3
# smtplib is built-in, no need for external package
aiofiles==23.2.0
pydantic==2.5.0
email-validator==2.1.0
google-api-python-client==2.108.0
